renamed entity to model and session to controller
