+++ recruitmentapp/src/main/java/se/kth/ict/iv1201/recruitmentapp/controller/RoleFacade.java
@@ -4,12 +4,12 @@
  * and open the template in the editor.
  */
 
-package se.kth.ict.iv1201.recruitmentapp.session;
+package se.kth.ict.iv1201.recruitmentapp.controller;
 
 import javax.ejb.Stateless;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
-import se.kth.ict.iv1201.recruitmentapp.entity.Role;
+import se.kth.ict.iv1201.recruitmentapp.model.Role;
 
 /**
  *
