+++ recruitmentapp/src/main/java/se/kth/ict/iv1201/recruitmentapp/model/Person.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package se.kth.ict.iv1201.recruitmentapp.entity;
+package se.kth.ict.iv1201.recruitmentapp.model;
 
 import java.io.Serializable;
 import java.util.Collection;
